распределил логику urlstorage по разным файлам в пределах одного модуля
